file rename: live-android-service-setup-business-logic-factory.ts -> android-service-configurator-factory.ts
